% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_binary_metrics}
\alias{calculate_binary_metrics}
\alias{calculate_binary_metrics,data.frame-method}
\title{Calculate Binary Classification Metrics}
\usage{
calculate_binary_metrics(aligned_df, ...)

\S4method{calculate_binary_metrics}{data.frame}(
  aligned_df,
  group_vars = NULL,
  total_comparisons = NULL
)
}
\arguments{
\item{aligned_df}{An object containing aligned counts.}

\item{...}{Additional arguments passed to methods.}

\item{group_vars}{A character vector of column names to group by.}

\item{total_comparisons}{An optional integer for the total number of comparisons.}
}
\description{
Calculate Binary Classification Metrics
}
\examples{
\dontrun{
# For Erin's ice seal survey, to compare raw vs. validated data
erin_raw_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_ir_detections.csv", package = "Optics")
raw_detections <- read_viame_csv(erin_raw_csv)

erin_truth_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_ir_detections_validated.csv", package = "Optics")
truth_detections <- read_viame_csv(erin_truth_csv)

raw_counts <- calculate_maxn(raw_detections)
truth_counts <- calculate_maxn(truth_detections)

aligned_df <- align_counts(
 raw_counts, 
 truth_counts,
 by = c("video_id", "category_name"),
 model_col = maxn,
 truth_col = maxn
)

binary_metrics <- calculate_binary_metrics(aligned_df)
}
}
