% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_confusion_matrix}
\alias{calculate_confusion_matrix}
\alias{calculate_confusion_matrix,data.frame-method}
\title{The rest of the functions are utilities that operate on generic data.frames.
They are converted to S4 generics for consistency.}
\usage{
calculate_confusion_matrix(aligned_df, ...)

\S4method{calculate_confusion_matrix}{data.frame}(
  aligned_df,
  group_vars,
  species_col = Species,
  model_col = model_count,
  truth_col = truth_count
)
}
\description{
The rest of the functions are utilities that operate on generic data.frames.
They are converted to S4 generics for consistency.
}
