% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_maxn}
\alias{calculate_maxn}
\alias{calculate_maxn,OpticsDetections-method}
\alias{calculate_maxn,data.frame-method}
\title{Calculate MaxN (Maximum Number of Individuals)}
\usage{
calculate_maxn(object, ...)

\S4method{calculate_maxn}{OpticsDetections}(object, group_cols = NULL)

\S4method{calculate_maxn}{data.frame}(object, group_cols = NULL)
}
\arguments{
\item{object}{An object containing detection data.}

\item{...}{Additional arguments passed to methods.}

\item{group_cols}{A character vector of additional column names to group by.}
}
\value{
A \code{tibble} with columns for each grouping variable and a \code{maxn}
column containing the calculated MaxN value.
}
\description{
Calculates the maximum number of individuals of each species category
observed in any single frame within each video. This is a standard metric
for baited remote underwater video (BRUV) surveys.
}
\examples{
# Create a sample OpticsDetections object
sample_df <- dplyr::tibble(
  video_id = c(rep("video1", 4), rep("video2", 3)),
  frame_index = c(1, 1, 2, 2, 1, 1, 1),
  category_name = c("FishA", "FishA", "FishA", "FishB", "FishA", "FishA", "FishA")
)
detections_obj <- OpticsDetections(
  data = sample_df,
  source_file = "dummy_file.csv",
  ingest_format = "dummy"
)

# Calculate MaxN for each video and species
calculate_maxn(detections_obj)

\dontrun{
# For Erin's ice seal survey, calculate MaxN by camera view
erin_csv_c <- system.file("extdata", 
"ice_seals_2025_fl223_C_rgb_irDetectionsTransposed_processed.csv", package = "Optics")
detections_c <- read_viame_csv(erin_csv_c, video_id = "center")

erin_csv_l <- system.file("extdata", 
"ice_seals_2025_fl223_L_rgb_irDetectionsTransposed_processed.csv", package = "Optics")
detections_l <- read_viame_csv(erin_csv_l, video_id = "left")

erin_csv_r <- system.file("extdata", 
"ice_seals_2025_fl223_R_rgb_irDetectionsTransposed_processed.csv", package = "Optics")
detections_r <- read_viame_csv(erin_csv_r, video_id = "right")

all_detections <- rbind(detections_c@data, detections_l@data, detections_r@data)

maxn <- calculate_maxn(all_detections, group_cols = "video_id")
}

}
