% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{OpticsDetections}
\alias{OpticsDetections}
\title{Constructor for OpticsDetections objects}
\usage{
OpticsDetections(data, source_file, ingest_format)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} of detection data.}

\item{source_file}{The path to the source data file.}

\item{ingest_format}{The format of the ingested data (e.g., "kwcoco").}
}
\value{
An \code{OpticsDetections} object.
}
\description{
A user-friendly constructor to create instances of the \code{OpticsDetections}
class. It wraps the standard \code{new()} method and provides a more intuitive
interface.
}
