% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_multiclass_confusion_matrix}
\alias{plot_multiclass_confusion_matrix}
\alias{plot_multiclass_confusion_matrix,data.frame-method}
\title{Plot a Multi-Class Confusion Matrix Heatmap}
\usage{
plot_multiclass_confusion_matrix(confusion_df, ...)

\S4method{plot_multiclass_confusion_matrix}{data.frame}(confusion_df, title = "Confusion Matrix")
}
\arguments{
\item{confusion_df}{A data frame from \code{calculate_confusion_matrix()}.}

\item{...}{Additional arguments.}
}
\description{
Plot a Multi-Class Confusion Matrix Heatmap
}
