% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\alias{plot_roc_curve,data.frame-method}
\title{Plot a Receiver Operating Characteristic (ROC) Curve}
\usage{
plot_roc_curve(detection_df, ...)

\S4method{plot_roc_curve}{data.frame}(detection_df, title = "ROC Curve", model_col = NULL)
}
\arguments{
\item{detection_df}{A data frame of classified detections.}

\item{...}{Additional arguments.}
}
\description{
Plot a Receiver Operating Characteristic (ROC) Curve
}
\examples{
\dontrun{
# For Abi's AUV data
kwcoco_file <- system.file("extdata", "AUV_viame_test_detections.coco.json", package = "Optics")
detections <- read_kwcoco(kwcoco_file)

# This is a placeholder for truth data
truth_detections <- detections 

classified_detections <- classify_detections(
raw_detections = detections@data,
validated_detections = truth_detections@data,
detection_id = annotation_id
)

plot_roc_curve(classified_detections)
}
}
