% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_scalpred_pr_curve}
\alias{plot_scalpred_pr_curve}
\alias{plot_scalpred_pr_curve,data.frame-method}
\title{Plot a ScalPred Precision-Recall Curve}
\usage{
plot_scalpred_pr_curve(summary_df, ...)

\S4method{plot_scalpred_pr_curve}{data.frame}(
  summary_df,
  model_col = NULL,
  title = "ScalPred Precision-Recall Curve"
)
}
\arguments{
\item{summary_df}{A data frame from \code{calculate_scalpred_metrics()}.}

\item{...}{Additional arguments.}

\item{model_col}{Optional unquoted model column for faceting.}

\item{title}{Plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a ScalPred Precision-Recall Curve
}
