% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{analyze_performance_drivers}
\alias{analyze_performance_drivers}
\alias{analyze_performance_drivers,data.frame-method}
\title{Analyze Performance Drivers}
\usage{
analyze_performance_drivers(aligned_df, ...)

\S4method{analyze_performance_drivers}{data.frame}(aligned_df, group_vars)
}
\description{
Analyze Performance Drivers
}
\examples{
\dontrun{
# For Tom & Michael's coral survey
aligned_df <- dplyr::tibble(
 video_id = "v1", Species = "Acropora", model_count = 5, truth_count = 10,
 n_species_truth = 1, total_individuals_truth = 10
)
performance_drivers <- analyze_performance_drivers(aligned_df, 
group_vars = c("video_id", "category_name"))
}
}
