% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{analyze_performance_drivers}
\alias{analyze_performance_drivers}
\alias{analyze_performance_drivers,data.frame-method}
\title{Analyze Performance Drivers}
\usage{
analyze_performance_drivers(aligned_df, ...)

\S4method{analyze_performance_drivers}{data.frame}(aligned_df, group_vars)
}
\description{
Analyze Performance Drivers
}
\examples{
\dontrun{
# For Tom & Michael's coral survey
# Assuming 'aligned_df' is created from align_counts()
performance_drivers <- analyze_performance_drivers(aligned_df, 
group_vars = c("video_id", "category_name"))
}
}
