% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{analyze_reviewer_effort}
\alias{analyze_reviewer_effort}
\alias{analyze_reviewer_effort,data.frame,data.frame-method}
\title{Analyze Reviewer Effort}
\usage{
analyze_reviewer_effort(raw_df, validated_df, ...)

\S4method{analyze_reviewer_effort}{data.frame,data.frame}(raw_df, validated_df, group_vars = NULL)
}
\description{
Analyze Reviewer Effort
}
\examples{
# Example using Erin's ice seal data
raw_csv_data <- c("1,video1,10,100,100,200,200,1,0.95,"ringed_seal",1")
validated_csv_data <- c("1,video1,10,100,100,200,200,1,1.0,"harbor_seal",1")
raw_csv_path <- tempfile(fileext = ".csv")
validated_csv_path <- tempfile(fileext = ".csv")
writeLines(c("# h1", "# h2", raw_csv_data), raw_csv_path)
writeLines(c("# h1", "# h2", validated_csv_data), validated_csv_path)

raw_df <- read_viame_csv(raw_csv_path)@data
validated_df <- read_viame_csv(validated_csv_path)@data

effort_analysis <- analyze_reviewer_effort(raw_df, validated_df)
print(effort_analysis)

unlink(raw_csv_path)
unlink(validated_csv_path)
}
