% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{analyze_reviewer_effort}
\alias{analyze_reviewer_effort}
\alias{analyze_reviewer_effort,data.frame,data.frame-method}
\title{Analyze Reviewer Effort}
\usage{
analyze_reviewer_effort(raw_df, validated_df, ...)

\S4method{analyze_reviewer_effort}{data.frame,data.frame}(raw_df, validated_df, group_vars = NULL)
}
\description{
Analyze Reviewer Effort
}
