% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{analyze_reviewer_effort}
\alias{analyze_reviewer_effort}
\alias{analyze_reviewer_effort,data.frame,data.frame-method}
\title{Analyze Reviewer Effort}
\usage{
analyze_reviewer_effort(raw_df, validated_df, ...)

\S4method{analyze_reviewer_effort}{data.frame,data.frame}(raw_df, validated_df, group_vars = NULL)
}
\description{
Analyze Reviewer Effort
}
\examples{
\dontrun{
# For Erin's ice seal survey
erin_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_rgb_irDetectionsTransposed_processed.csv", package = "Optics")
raw_df <- read_viame_csv(erin_csv)@data

erin_truth_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_ir_detections_validated.csv", package = "Optics")
validated_df <- read_viame_csv(erin_truth_csv)@data

effort_analysis <- analyze_reviewer_effort(raw_df, validated_df)
}
}
