% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_binary_metrics}
\alias{calculate_binary_metrics}
\alias{calculate_binary_metrics,data.frame-method}
\title{Calculate Binary Classification Metrics}
\usage{
calculate_binary_metrics(aligned_df, ...)

\S4method{calculate_binary_metrics}{data.frame}(
  aligned_df,
  group_vars = NULL,
  total_comparisons = NULL
)
}
\arguments{
\item{aligned_df}{An object containing aligned counts.}

\item{...}{Additional arguments passed to methods.}

\item{group_vars}{A character vector of column names to group by.}

\item{total_comparisons}{An optional integer for the total number of comparisons.}
}
\description{
Calculate Binary Classification Metrics
}
