% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{classify_detections}
\alias{classify_detections}
\alias{classify_detections,data.frame,data.frame-method}
\title{Classify Detections as True/False Positives}
\usage{
classify_detections(raw_detections, validated_detections, ...)

\S4method{classify_detections}{data.frame,data.frame}(raw_detections, validated_detections, detection_id)
}
\arguments{
\item{raw_detections}{An object of raw detections.}

\item{validated_detections}{An object of validated detections.}

\item{...}{Additional arguments.}

\item{detection_id}{The unquoted column name for the unique detection identifier.}
}
\description{
Classify Detections as True/False Positives
}
\examples{
\dontrun{
# For Erin's ice seal survey
erin_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_rgb_irDetectionsTransposed_processed.csv", package = "Optics")
raw_detections <- read_viame_csv(erin_csv)

erin_truth_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_ir_detections_validated.csv", package = "Optics")
validated_detections <- read_viame_csv(erin_truth_csv)

classified_detections <- classify_detections(
raw_detections = raw_detections@data,
validated_detections = validated_detections@data,
detection_id = annotation_id
)
}
}
