% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{classify_detections}
\alias{classify_detections}
\alias{classify_detections,data.frame,data.frame-method}
\title{Classify Detections as True/False Positives}
\usage{
classify_detections(raw_detections, validated_detections, ...)

\S4method{classify_detections}{data.frame,data.frame}(raw_detections, validated_detections, detection_id)
}
\arguments{
\item{raw_detections}{An object of raw detections.}

\item{validated_detections}{An object of validated detections.}

\item{...}{Additional arguments.}

\item{detection_id}{The unquoted column name for the unique detection identifier.}
}
\description{
Classify Detections as True/False Positives
}
\examples{
# Example using Erin's ice seal data.
model_csv_data <- c("1,video1,10,100,100,200,200,1,0.95,"ringed_seal",1")
truth_csv_data <- c("2,video1,10,110,110,210,210,1,1.0,"ringed_seal",1")
model_csv_path <- tempfile(fileext = ".csv")
truth_csv_path <- tempfile(fileext = ".csv")
writeLines(c("# h1", "# h2", model_csv_data), model_csv_path)
writeLines(c("# h1", "# h2", truth_csv_data), truth_csv_path)

raw_detections <- read_viame_csv(model_csv_path)
validated_detections <- read_viame_csv(truth_csv_path)

classified_detections <- classify_detections(
  raw_detections = raw_detections@data,
  validated_detections = validated_detections@data,
  detection_id = annotation_id
)
print(classified_detections)

unlink(model_csv_path)
unlink(truth_csv_path)
}
