% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{classify_detections}
\alias{classify_detections}
\alias{classify_detections,data.frame,data.frame-method}
\title{Classify Detections as True/False Positives}
\usage{
classify_detections(raw_detections, validated_detections, ...)

\S4method{classify_detections}{data.frame,data.frame}(raw_detections, validated_detections, detection_id)
}
\arguments{
\item{raw_detections}{An object of raw detections.}

\item{validated_detections}{An object of validated detections.}

\item{...}{Additional arguments.}

\item{detection_id}{The unquoted column name for the unique detection identifier.}
}
\description{
Classify Detections as True/False Positives
}
