% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{get_disagreement_report}
\alias{get_disagreement_report}
\alias{get_disagreement_report,data.frame-method}
\title{Get a Report of Disagreements}
\usage{
get_disagreement_report(aligned_df, ...)

\S4method{get_disagreement_report}{data.frame}(aligned_df, group_vars, top_n = 10)
}
\description{
Get a Report of Disagreements
}
