% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\alias{plot_bland_altman,data.frame-method}
\title{Plot a Bland-Altman Plot}
\usage{
plot_bland_altman(aligned_df, ...)

\S4method{plot_bland_altman}{data.frame}(
  aligned_df,
  title = "Bland-Altman Agreement Plot",
  model_col = NULL
)
}
\arguments{
\item{aligned_df}{A data frame of aligned counts.}

\item{...}{Additional arguments.}
}
\description{
Plot a Bland-Altman Plot
}
\examples{
\dontrun{
erin_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_rgb_irDetectionsTransposed_processed.csv", package = "Optics")
model_detections <- read_viame_csv(erin_csv)

erin_truth_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_ir_detections_validated.csv", package = "Optics")
truth_detections <- read_viame_csv(erin_truth_csv)

model_counts <- calculate_maxn(model_detections)
truth_counts <- calculate_maxn(truth_detections)

aligned_df <- align_counts(
 model_counts, 
 truth_counts,
 by = c("video_id", "category_name"),
 model_col = maxn,
 truth_col = maxn
 )

plot_bland_altman(aligned_df)
}
}
