% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\alias{plot_bland_altman,data.frame-method}
\title{Plot a Bland-Altman Plot}
\usage{
plot_bland_altman(aligned_df, ...)

\S4method{plot_bland_altman}{data.frame}(
  aligned_df,
  title = "Bland-Altman Agreement Plot",
  model_col = NULL
)
}
\arguments{
\item{aligned_df}{A data frame of aligned counts.}

\item{...}{Additional arguments.}
}
\description{
Plot a Bland-Altman Plot
}
