% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\alias{plot_bland_altman,data.frame-method}
\title{Plot a Bland-Altman Plot}
\usage{
plot_bland_altman(aligned_df, ...)

\S4method{plot_bland_altman}{data.frame}(
  aligned_df,
  title = "Bland-Altman Agreement Plot",
  model_col = NULL
)
}
\arguments{
\item{aligned_df}{A data frame of aligned counts.}

\item{...}{Additional arguments.}

\item{title}{Character. Plot title.}

\item{model_col}{Symbol or character for model column.}
}
\description{
Plot a Bland-Altman Plot
}
\examples{
# Example using Erin's ice seal data.
model_csv_data <- c("1,video1,10,100,100,200,200,1,0.95,\"ringed_seal\",1")
truth_csv_data <- c("1,video1,10,100,100,200,200,1,1.0,\"ringed_seal\",1")
model_csv_path <- tempfile(fileext = ".csv")
truth_csv_path <- tempfile(fileext = ".csv")
writeLines(c("# h1", "# h2", model_csv_data), model_csv_path)
writeLines(c("# h1", "# h2", truth_csv_data), truth_csv_path)

model_detections <- read_viame_csv(model_csv_path)
truth_detections <- read_viame_csv(truth_csv_path)

aligned_df <- align_counts(
 calculate_maxn(model_detections),
 calculate_maxn(truth_detections),
 by = c("video_id", "category_name")
)

plot_bland_altman(aligned_df)

unlink(model_csv_path)
unlink(truth_csv_path)
}
