% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\alias{plot_confusion_matrix,data.frame-method}
\title{Plot a Confusion Matrix}
\usage{
plot_confusion_matrix(metrics_df, ...)

\S4method{plot_confusion_matrix}{data.frame}(metrics_df, title = "Confusion Matrix", model_col = NULL)
}
\arguments{
\item{metrics_df}{A data frame of metrics.}

\item{...}{Additional arguments.}
}
\description{
Plot a Confusion Matrix
}
