% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\alias{plot_confusion_matrix,data.frame-method}
\title{Plot a Confusion Matrix}
\usage{
plot_confusion_matrix(metrics_df, ...)

\S4method{plot_confusion_matrix}{data.frame}(metrics_df, title = "Confusion Matrix", model_col = NULL)
}
\arguments{
\item{metrics_df}{A data frame of metrics.}

\item{...}{Additional arguments.}
}
\description{
Plot a Confusion Matrix
}
\examples{
\dontrun{
erin_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_rgb_irDetectionsTransposed_processed.csv", package = "Optics")
model_detections <- read_viame_csv(erin_csv)

erin_truth_csv <- system.file("extdata", 
"ice_seals_2025_fl223_C_ir_detections_validated.csv", package = "Optics")
truth_detections <- read_viame_csv(erin_truth_csv)

model_counts <- calculate_maxn(model_detections)
truth_counts <- calculate_maxn(truth_detections)

aligned_df <- align_counts(
 model_counts, 
 truth_counts,
 by = c("video_id", "category_name"),
 model_col = maxn,
 truth_col = maxn
 )

metrics_df <- calculate_binary_metrics(aligned_df)

plot_confusion_matrix(metrics_df)
}
}
