% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\alias{plot_confusion_matrix,data.frame-method}
\title{Plot a Confusion Matrix}
\usage{
plot_confusion_matrix(metrics_df, ...)

\S4method{plot_confusion_matrix}{data.frame}(metrics_df, title = "Confusion Matrix", model_col = NULL)
}
\arguments{
\item{metrics_df}{A data frame of metrics.}

\item{...}{Additional arguments.}
}
\description{
Plot a Confusion Matrix
}
\examples{
# Example using Erin's ice seal data.
model_csv_data <- c("1,video1,10,100,100,200,200,1,0.95,"ringed_seal",1")
truth_csv_data <- c("1,video1,10,100,100,200,200,1,1.0,"ringed_seal",1")
model_csv_path <- tempfile(fileext = ".csv")
truth_csv_path <- tempfile(fileext = ".csv")
writeLines(c("# h1", "# h2", model_csv_data), model_csv_path)
writeLines(c("# h1", "# h2", truth_csv_data), truth_csv_path)

model_detections <- read_viame_csv(model_csv_path)
truth_detections <- read_viame_csv(truth_csv_path)

aligned_df <- align_counts(
 calculate_maxn(model_detections),
 calculate_maxn(truth_detections),
 by = c("video_id", "category_name")
)

metrics_df <- calculate_binary_metrics(aligned_df, total_comparisons = 2)

plot_confusion_matrix(metrics_df)

unlink(model_csv_path)
unlink(truth_csv_path)
}
