% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_counts_scatterplot}
\alias{plot_counts_scatterplot}
\alias{plot_counts_scatterplot,data.frame-method}
\title{Plot a Scatterplot of Model vs. Truth Counts}
\usage{
plot_counts_scatterplot(aligned_df, ...)

\S4method{plot_counts_scatterplot}{data.frame}(
  aligned_df,
  title = "Model vs. Truth Counts",
  model_col = NULL
)
}
\arguments{
\item{aligned_df}{A data frame containing aligned counts.}

\item{...}{Additional arguments.}

\item{title}{An optional title for the plot.}

\item{model_col}{An optional unquoted column name for faceting by model.}
}
\description{
Plot a Scatterplot of Model vs. Truth Counts
}
