% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_counts_scatterplot}
\alias{plot_counts_scatterplot}
\alias{plot_counts_scatterplot,data.frame-method}
\title{Plot a Scatterplot of Model vs. Truth Counts}
\usage{
plot_counts_scatterplot(aligned_df, ...)

\S4method{plot_counts_scatterplot}{data.frame}(
  aligned_df,
  title = "Model vs. Truth Counts",
  model_col = NULL
)
}
\arguments{
\item{aligned_df}{A data frame containing aligned counts.}

\item{...}{Additional arguments.}

\item{title}{An optional title for the plot.}

\item{model_col}{An optional unquoted column name for faceting by model.}
}
\description{
Plot a Scatterplot of Model vs. Truth Counts
}
\examples{
# Example using GFISHER data.
model_csv_data <- c("1,video1,10,100,100,200,200,1,0.95,\"Gadus morhua\",1")
truth_csv_data <- c("1,video1,10,100,100,200,200,1,1.0,\"Gadus morhua\",1")
model_csv_path <- tempfile(fileext = ".csv")
truth_csv_path <- tempfile(fileext = ".csv")
writeLines(c("# h1", "# h2", model_csv_data), model_csv_path)
writeLines(c("# h1", "# h2", truth_csv_data), truth_csv_path)

model_detections <- read_viame_csv(model_csv_path)
truth_detections <- read_viame_csv(truth_csv_path)

aligned_df <- align_counts(
 calculate_maxn(model_detections), 
 calculate_maxn(truth_detections),
 by = c("video_id", "category_name")
 )

plot_counts_scatterplot(aligned_df)

unlink(model_csv_path)
unlink(truth_csv_path)
}
