% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_multiclass_confusion_matrix}
\alias{plot_multiclass_confusion_matrix}
\alias{plot_multiclass_confusion_matrix,data.frame-method}
\title{Plot a Multi-Class Confusion Matrix Heatmap}
\usage{
plot_multiclass_confusion_matrix(confusion_df, ...)

\S4method{plot_multiclass_confusion_matrix}{data.frame}(confusion_df, title = "Confusion Matrix")
}
\arguments{
\item{confusion_df}{A data frame from \code{calculate_confusion_matrix()}.}

\item{...}{Additional arguments.}
}
\description{
Plot a Multi-Class Confusion Matrix Heatmap
}
\examples{
\dontrun{
# For Tom & Michael's coral survey
kwcoco_file <- system.file("extdata", "AUV_viame_test_detections.coco.json", package = "Optics")
detections <- read_kwcoco(kwcoco_file)

# This is a placeholder for truth data
truth_detections <- detections 

model_counts <- calculate_maxn(detections)
truth_counts <- calculate_maxn(truth_detections)

aligned_df <- align_counts(
 model_counts, 
 truth_counts,
 by = c("video_id", "category_name"),
 model_col = maxn,
 truth_col = maxn
 )

confusion_df <- calculate_confusion_matrix(aligned_df, 
group_vars = "video_id", species_col = category_name)

plot_multiclass_confusion_matrix(confusion_df)
}
}
