% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_multiclass_confusion_matrix}
\alias{plot_multiclass_confusion_matrix}
\alias{plot_multiclass_confusion_matrix,data.frame-method}
\title{Plot a Multi-Class Confusion Matrix Heatmap}
\usage{
plot_multiclass_confusion_matrix(confusion_df, ...)

\S4method{plot_multiclass_confusion_matrix}{data.frame}(confusion_df, title = "Confusion Matrix")
}
\arguments{
\item{confusion_df}{A data frame from \code{calculate_confusion_matrix()}.}

\item{...}{Additional arguments.}
}
\description{
Plot a Multi-Class Confusion Matrix Heatmap
}
\examples{
# Example using Tom & Michael's coral survey data.
model_csv_data <- c("1,video1,10,100,100,200,200,1,0.95,"Acropora",1")
truth_csv_data <- c("1,video1,10,100,100,200,200,1,1.0,"Pocillopora",1")
model_csv_path <- tempfile(fileext = ".csv")
truth_csv_path <- tempfile(fileext = ".csv")
writeLines(c("# h1", "# h2", model_csv_data), model_csv_path)
writeLines(c("# h1", "# h2", truth_csv_data), truth_csv_path)

model_detections <- read_viame_csv(model_csv_path)
truth_detections <- read_viame_csv(truth_csv_path)

aligned_df <- align_counts(
 calculate_maxn(model_detections),
 calculate_maxn(truth_detections),
 by = c("video_id", "category_name")
)

confusion_df <- calculate_confusion_matrix(aligned_df,
 group_vars = "video_id", species_col = category_name)

plot_multiclass_confusion_matrix(confusion_df)

unlink(model_csv_path)
unlink(truth_csv_path)
}
