% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_pr_curve}
\alias{plot_pr_curve}
\alias{plot_pr_curve,data.frame-method}
\title{Plot a Precision-Recall (PR) Curve}
\usage{
plot_pr_curve(detection_df, ...)

\S4method{plot_pr_curve}{data.frame}(detection_df, model_col = NULL, title = "Precision-Recall Curve")
}
\arguments{
\item{detection_df}{A data frame of classified detections.}

\item{...}{Additional arguments.}
}
\description{
Plot a Precision-Recall (PR) Curve
}
