% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\alias{plot_roc_curve,data.frame-method}
\title{Plot a Receiver Operating Characteristic (ROC) Curve}
\usage{
plot_roc_curve(detection_df, ...)

\S4method{plot_roc_curve}{data.frame}(detection_df, title = "ROC Curve", model_col = NULL)
}
\arguments{
\item{detection_df}{A data frame of classified detections.}

\item{...}{Additional arguments.}

\item{title}{Character. Plot title.}

\item{model_col}{Symbol or character for model column.}
}
\description{
Plot a Receiver Operating Characteristic (ROC) Curve
}
\examples{
# Example using Abi's AUV data.
raw_csv_data <- c("1,video1,10,100,100,200,200,1,0.95,\"sea_star\",1")
validated_csv_data <- c("1,video1,10,100,100,200,200,1,1.0,\"sea_star\",1")
raw_csv_path <- tempfile(fileext = ".csv")
validated_csv_path <- tempfile(fileext = ".csv")
writeLines(c("# h1", "# h2", raw_csv_data), raw_csv_path)
writeLines(c("# h1", "# h2", validated_csv_data), validated_csv_path)

raw_detections <- read_viame_csv(raw_csv_path)
validated_detections <- read_viame_csv(validated_csv_path)

classified_detections <- classify_detections(
  raw_detections = raw_detections@data,
  validated_detections = validated_detections@data,
  detection_id = annotation_id
)

plot_roc_curve(classified_detections)

unlink(raw_csv_path)
unlink(validated_csv_path)
}
